change refCounter to UINT32
git-svn-id: http://192.168.20.11/sequoiadb/trunk@38125 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/include/dpsTransLRB.hpp
@@ -58,7 +58,7 @@ namespace engine
       UTIL_OBJIDX eduLrbIdxPrev ;   // idx of prev LRB in chain EDU owning in tx
       UTIL_OBJIDX lrbHdrIdx ;       // idx of the LRB Header
       UTIL_OBJIDX nextLRBIdx ;      // idx of next LRB in the owner/waiter chain
-      UINT64 refCounter ;           // lock reference counter
+      UINT32 refCounter ;           // lock reference counter
       DPS_TRANSLOCK_TYPE lockMode ; // lock mode, UINT8, 1 byte
       UINT8 pad[3] ;                // 3 byte for padding
    } ;
